SVY-20262 check the rhino 1.8 unit tests
for some reason in compiled mode the literal number 123 is a integer and
not a double (but in interpreted mode it is a double)
Not sure why that is because Servoy doesn't really use compiled mode

+++ rhino/src/test/java/org/mozilla/javascript/tests/NativeJavaObjectTest.java
@@ -37,8 +37,8 @@ public class NativeJavaObjectTest {
                 Object result = cx.evaluateString(scope, source, "source", 1, null);
                 assertTrue(result instanceof NativeJavaObject);
                 Object rawObj = ((NativeJavaObject) result).unwrap();
-                assertTrue(rawObj instanceof Double);
-                assertEquals(Double.valueOf(123), rawObj);
+                assertTrue(rawObj instanceof Integer); // servoy patched test
+                assertEquals(Integer.valueOf(123), rawObj); 
             }
 
             {
